Pass the view as this when adding a single element during initial comments population
When initially populating the TopComments collection, the call to fetch
triggers reset, which doesn't trigger multiple add events and we need
to call addOne by ourselves.

However, we need to make sure that `this` is the right object when we
call addOne, because otherwise it's window, which messes up this.el and
this.$ isn't restricted to the view element only.