Simplify generating data for the comment filters
Fetch all comments from the database only once and give them to the
functions which gather the data for all the filters.

A lot simpler method for getting all the paths – for each comment get
the directory up to the 2nd level and then filter the uniques and not
empty ones.

Don't add % in the end of the paths. It was annoying to look at.

Don't let arbitrary data in search queries – escape IN query values,
limit column names to existing ones.

Get rid of generic LIKE search queries. Limit ourselves to a prefix
query.