Don't include a value for the auto-increment for Postgres
When inserting a comment we used to pass NULL as the value for the
auto-increment id, but this isn't valid syntax for Postgres.

Instead, we are explicitly sending the names of the columns we are
inserting and we skip the id. This way it should work on both SQLite
and Postgres.

Fixes #4.