Fix line numbering display issues when commenting on a line of a changeset
To associate a comment with a specific line of the changeset, we use the row number in the diff table displayed by Trac.  However, this didn't work for diffs with multiple files (if the first file had five lines, the second file's first line would be line #6) or deleted lines.

I added a displayLine variable that controls what we show to the user: the actual line number of the file, or its former line number if the line was deleted.  (We still use the absolute row number to store the comment, since otherwise we couldn't disambiguate which line the comment belongs to.)