Fix "Path" display in Code Comments table for changeset comments
Be more specific about when a comment is on a changeset vs. a file.  Also, remove the line number from the changeset for now, because it's confusing (it refers to the absolute line number of the changeset, NOT the line within the file).  To show that a comment refers to a particular line in a changeset, we add "(in filename)".