Use a single template for line comments
Instead of using two different templates with a minor difference, use one template with an argument that tells it whether to expand the TH to one column or two.

This uses a templateData attribute on the view; we add an argument to it on initialize() and pass it to the template on render().