Hack to get anchors to comments working for changeset view
Just adding new anchors and permalinks isn't good enough; the Backbone view hasn't rendered the comments yet, so jumping to the anchor fails.

This is a hack that hooks up to _.defer and uses a setTimeout, because even _.defer is not waiting long enough to make sure the DOM has rendered before it runs.