Send notifications when comments are created
Added a new interface - ICodeCommentChangeListener - to use for
notifications.

Added a new component - CodeCommentSystem - to act as the extension
point for the above interface. (This could have been added to one of the
other components, but I have plans for other refactorings and wanted a
clean base for them.)

Extended Comments.create() to trigger comment_created events.

Added a new component - CodeCommentChangeListener - to respond to
comment_created events.

Sub-classed trac.notification.NotifyEmail to generate notification
emails. Notifications are sent to the author of the resouce being
commented on, and anyone else who has commented on the same resource.
There is additional logic to determine who the notification should be
*sent* and *copied* to.

Added an option to control whether notifications are sent to the person
making the comment (off by default).

Removed unused variable.