Refactored `CodeCommentNotifyEmail` to use subscriptions
Added a `for_comment` query helper.

Changed the notification scheme so that resource authors **always**
receive a "to" notification.

Changed the `notify` column to be a boolean.

Changed the `for_*` signatures to accept an optional `notify` flag.

Changed the `select` signature to accept an optional `notify` flag, and
the method to include the flag in the query.

Changed the `select` method to handle `dict`, `list, and `bool` query
arguments.