Move page-level comments to the bottom
If they’re at the top and somebody goes to the page just to look at the
file and not at the comments, a huge discussion at the top can be very
distracting.

We assume that people will go to pages more often to see the
files/diffs/commits than to look at comments.